fix: cast MoE weighted_out to input dtype before scatter_add_ in 3B/8B/70B
The MoEGate uses .float() for null-expert masking, which promotes
topk_weight to fp32. When multiplied with sorted_out (bf16), the
result is fp32 but routed_out is bf16, causing scatter_add_ to fail
with "Expected self.dtype to be equal to src.dtype".

+++ experiments/tests/Test_9_extend_test8_to_3000steps_save_state/code/src/models/recurrence_model_70b.py
@@ -1720,7 +1720,7 @@ class MoEFFN(nn.Module):
         else:
             sorted_out = torch.empty(0, D, device=device, dtype=dtype)
 
-        weighted_out = sorted_out * sorted_weights.unsqueeze(-1)
+        weighted_out = (sorted_out * sorted_weights.unsqueeze(-1)).to(dtype)
         routed_out = torch.zeros(N, D, device=device, dtype=dtype)
         routed_out.scatter_add_(
             0, sorted_token_indices.unsqueeze(-1).expand(-1, D), weighted_out
